statistical_test
A statistical test is a formal test of a hypothesis, in which data from a sample is used to assess whether or not a particular hypothesis about a population is true. Statistical tests are used to evaluate the validity of a hypothesis, and to determine the strength of evidence for or against the hypothesis. They are also used to compare two or more populations in order to determine if the difference between them is significant.