EM
Expectation Maximization (EM) is an iterative algorithm used to find maximum likelihood parameters of a model given observed data. It is especially useful in cases where the data contains hidden or latent variables. The algorithm works by first initializing the model parameters, then alternating between two steps: expectation and maximization. The expectation step calculates the expected value of the latent variables given the current parameters, and the maximization step updates the parameters of the model to maximize the expected value of the latent variables. Once the algorithm converges, the parameters of the model will be the maximum likelihood estimates.