Markov chain
A Markov chain is a mathematical system that models transitions from one state to another. It can be used to find the probability of a particular sequence of events occurring. In particular, it is used to model systems with random processes, such as weather systems or stock prices. The most common application is to calculate the probability of a certain outcome given a certain set of inputs. For example, a Markov chain can be used to calculate the probability of a particular stock price going up or down based on the previous stock prices.